Cleaned up types
git-svn-id: https://svn.aros.org/svn/aros/trunk@28963 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/exec/rawdofmt.c
@@ -92,7 +92,7 @@
 #define PutCh(ch)                         \
 do                                        \
 {                                         \
-    switch ((ULONG)PutChProc)             \
+    switch (PutChProc)                    \
     {                                     \
     case RAWFMTFUNC_STRING:		  \
 	*UPutChData++ = ch;               \
